move delayed reveal to core/classes
