+++ docs/sdk.delayedreveal._constructor_.md
@@ -1,0 +1,21 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [DelayedReveal](./sdk.delayedreveal.md) &gt; [(constructor)](./sdk.delayedreveal._constructor_.md)
+
+## DelayedReveal.(constructor)
+
+Constructs a new instance of the `DelayedReveal` class
+
+<b>Signature:</b>
+
+```typescript
+constructor(contractWrapper: ContractWrapper<T>, storage: IStorage);
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  contractWrapper | ContractWrapper&lt;T&gt; |  |
+|  storage | [IStorage](./sdk.istorage.md) |  |
+
