+++ docs/sdk.delayedreveal.getbatchestoreveal.md
@@ -1,0 +1,17 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [DelayedReveal](./sdk.delayedreveal.md) &gt; [getBatchesToReveal](./sdk.delayedreveal.getbatchestoreveal.md)
+
+## DelayedReveal.getBatchesToReveal() method
+
+Gets the list of unrevealed NFT batches.
+
+<b>Signature:</b>
+
+```typescript
+getBatchesToReveal(): Promise<BatchToReveal[]>;
+```
+<b>Returns:</b>
+
+Promise&lt;[BatchToReveal](./sdk.batchtoreveal.md)<!-- -->\[\]&gt;
+
