+++ docs/sdk.delayedreveal.reveal.md
@@ -1,0 +1,25 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [DelayedReveal](./sdk.delayedreveal.md) &gt; [reveal](./sdk.delayedreveal.reveal.md)
+
+## DelayedReveal.reveal() method
+
+Reveal a batch of hidden NFTs
+
+<b>Signature:</b>
+
+```typescript
+reveal(batchId: BigNumberish, password: string): Promise<TransactionResult>;
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  batchId | BigNumberish | the id of the batch to reveal |
+|  password | string | the password |
+
+<b>Returns:</b>
+
+Promise&lt;[TransactionResult](./sdk.transactionresult.md)<!-- -->&gt;
+
