+++ docusaurus/docs/sdk.delayedreveal._constructor_.md
@@ -1,0 +1,25 @@
+---
+slug: /sdk.delayedreveal._constructor_
+title: DelayedReveal.(constructor)
+hide_title: true
+---
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[@thirdweb-dev/sdk](./sdk.md) &gt; [DelayedReveal](./sdk.delayedreveal.md) &gt; [(constructor)](./sdk.delayedreveal._constructor_.md)
+
+## DelayedReveal.(constructor)
+
+Constructs a new instance of the `DelayedReveal` class
+
+**Signature:**
+
+```typescript
+constructor(contractWrapper: ContractWrapper<T>, storage: IStorage);
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  contractWrapper | ContractWrapper&lt;T&gt; |  |
+|  storage | [IStorage](./sdk.istorage.md) |  |
