+++ docusaurus/docs/sdk.delayedreveal.createdelayedrevealbatch.md
@@ -1,0 +1,60 @@
+---
+slug: /sdk.delayedreveal.createdelayedrevealbatch
+title: DelayedReveal.createDelayedRevealBatch() method
+hide_title: true
+---
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[@thirdweb-dev/sdk](./sdk.md) &gt; [DelayedReveal](./sdk.delayedreveal.md) &gt; [createDelayedRevealBatch](./sdk.delayedreveal.createdelayedrevealbatch.md)
+
+## DelayedReveal.createDelayedRevealBatch() method
+
+Create a batch of encrypted NFTs that can be revealed at a later time.
+
+**Signature:**
+
+```typescript
+createDelayedRevealBatch(placeholder: NFTMetadataInput, metadatas: NFTMetadataInput[], password: string): Promise<TransactionResultWithId[]>;
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  placeholder | [NFTMetadataInput](./sdk.nftmetadatainput.md) | the placeholder NFT to show before the reveal |
+|  metadatas | [NFTMetadataInput](./sdk.nftmetadatainput.md)\[\] | the final NFTs that will be hidden |
+|  password | string | the password that will be used to reveal these NFTs |
+
+**Returns:**
+
+Promise&lt;[TransactionResultWithId](./sdk.transactionresultwithid.md)\[\]&gt;
+
+## Example
+
+
+```javascript
+// the real NFTs, these will be encrypted until your reveal them!
+const realNFTs = [{
+  name: "Common NFT #1",
+  description: "Common NFT, one of many.",
+  image: fs.readFileSync("path/to/image.png"),
+}, {
+  name: "Super Rare NFT #2",
+  description: "You got a Super Rare NFT!",
+  image: fs.readFileSync("path/to/image.png"),
+}];
+// A placeholder NFT that people will get immediately in their wallet, until the reveal happens!
+const placeholderNFT = {
+  name: "Hidden NFT",
+  description: "Will be revealed next week!"
+};
+// Create and encrypt the NFTs
+await contract.revealer.createDelayedRevealBatch(
+  placeholderNFT,
+  realNFTs,
+  "my secret password",
+);
+// Whenever you're ready, reveal your NFTs at any time!
+const batchId = 0; // the batch to reveal
+await contract.revealer.reveal(batchId, "my secret password");
+```
