+++ docusaurus/docs/sdk.delayedreveal.getbatchestoreveal.md
@@ -1,0 +1,21 @@
+---
+slug: /sdk.delayedreveal.getbatchestoreveal
+title: DelayedReveal.getBatchesToReveal() method
+hide_title: true
+---
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[@thirdweb-dev/sdk](./sdk.md) &gt; [DelayedReveal](./sdk.delayedreveal.md) &gt; [getBatchesToReveal](./sdk.delayedreveal.getbatchestoreveal.md)
+
+## DelayedReveal.getBatchesToReveal() method
+
+Gets the list of unrevealed NFT batches.
+
+**Signature:**
+
+```typescript
+getBatchesToReveal(): Promise<BatchToReveal[]>;
+```
+**Returns:**
+
+Promise&lt;[BatchToReveal](./sdk.batchtoreveal.md)\[\]&gt;
