+++ docusaurus/docs/sdk.delayedreveal.md
@@ -1,0 +1,32 @@
+---
+slug: /sdk.delayedreveal
+title: DelayedReveal class
+hide_title: true
+---
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[@thirdweb-dev/sdk](./sdk.md) &gt; [DelayedReveal](./sdk.delayedreveal.md)
+
+## DelayedReveal class
+
+Handles delayed reveal logic
+
+**Signature:**
+
+```typescript
+export declare class DelayedReveal<T extends DropERC721> 
+```
+
+## Constructors
+
+|  Constructor | Modifiers | Description |
+|  --- | --- | --- |
+|  [(constructor)(contractWrapper, storage)](./sdk.delayedreveal._constructor_.md) |  | Constructs a new instance of the <code>DelayedReveal</code> class |
+
+## Methods
+
+|  Method | Modifiers | Description |
+|  --- | --- | --- |
+|  [createDelayedRevealBatch(placeholder, metadatas, password)](./sdk.delayedreveal.createdelayedrevealbatch.md) |  | Create a batch of encrypted NFTs that can be revealed at a later time. |
+|  [getBatchesToReveal()](./sdk.delayedreveal.getbatchestoreveal.md) |  | Gets the list of unrevealed NFT batches. |
+|  [reveal(batchId, password)](./sdk.delayedreveal.reveal.md) |  | Reveal a batch of hidden NFTs |
