+++ docusaurus/docs/sdk.delayedreveal.reveal.md
@@ -1,0 +1,29 @@
+---
+slug: /sdk.delayedreveal.reveal
+title: DelayedReveal.reveal() method
+hide_title: true
+---
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[@thirdweb-dev/sdk](./sdk.md) &gt; [DelayedReveal](./sdk.delayedreveal.md) &gt; [reveal](./sdk.delayedreveal.reveal.md)
+
+## DelayedReveal.reveal() method
+
+Reveal a batch of hidden NFTs
+
+**Signature:**
+
+```typescript
+reveal(batchId: BigNumberish, password: string): Promise<TransactionResult>;
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  batchId | BigNumberish | the id of the batch to reveal |
+|  password | string | the password |
+
+**Returns:**
+
+Promise&lt;[TransactionResult](./sdk.transactionresult.md)&gt;
